full cov validators by height shard
